refactor: consolidate notes into card modules
