--- domain/notes.ts
@@ -1,5 +1,0 @@
-import { z } from 'zod';
-export const NOTES_MAX_LENGTH = 500;
-export const noteTextSchema = z.string().max(NOTES_MAX_LENGTH, {
-  error: `Note exceeds maximum length of ${NOTES_MAX_LENGTH} characters`,
-});
