+++ entrypoints/popup/components/notes/__tests__/useNoteEditor.test.tsx
@@ -1,7 +1,7 @@
 /** @vitest-environment happy-dom */
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { NOTES_MAX_LENGTH } from '@/domain/notes';
+import { NOTES_MAX_LENGTH } from '@/domain/cards';
 import { noteQueryKeys } from '@/entrypoints/popup/queries/notes';
 import { sendMessage } from '@/infrastructure/browser/messages';
 import { createMessageMock } from '@/test/utils/message-mocks';
